+++ incapsula/provider.go
@@ -6,7 +6,7 @@ import (
 )
 
 var baseURL string
-var incapRuleBaseURL string
+var apiV2BaseURL string
 var descriptions map[string]string
 
 func init() {
@@ -14,9 +14,9 @@ func init() {
 	// This endpoint is unlikely to change in the near future
 	baseURL = "https://my.incapsula.com/api/prov/v1"
 
-	// Efforts on APIv2 are underway and IncapRules are supported
-	// The other endpoints will eventually move over but we'll need this for now
-	incapRuleBaseURL = "https://my.imperva.com/api/prov/v2"
+	// Efforts on APIv2 are underway and newer resource are supported
+	// The other endpoints will eventually move over but we'll need the following for now
+	apiV2BaseURL = "https://my.imperva.com/api/prov/v2"
 
 	descriptions = map[string]string{
 		"api_id": "The API identifier for API operations. You can retrieve this\n" +
@@ -34,10 +34,10 @@ func configureProvider(d *schema.ResourceData) (interface{}, error) {
 	apiKey := d.Get("api_key").(string)
 
 	config := Config{
-		APIID:            apiID,
-		APIKey:           apiKey,
-		BaseURL:          baseURL,
-		IncapRuleBaseURL: incapRuleBaseURL,
+		APIID:        apiID,
+		APIKey:       apiKey,
+		BaseURL:      baseURL,
+		APIV2BaseURL: apiV2BaseURL,
 	}
 
 	return config.Client()
@@ -70,6 +70,7 @@ func Provider() terraform.ResourceProvider {
 			"incapsula_data_center":             resourceDataCenter(),
 			"incapsula_data_center_server":      resourceDataCenterServer(),
 			"incapsula_custom_certificate":      resourceCertificate(),
+			"incapsula_cache_rule":              resourceCacheRule(),
 		},
 
 		ConfigureFunc: configureProvider,